chore: Refactor scheduler handlers: Split into individual files and create a base handler class
- Removed the monolithic handlers.py file and created separate handler files for each scheduling operation.
- Introduced a base class `SchedulerHandler` to encapsulate common functionality and enforce a consistent interface across all handlers.
- Updated imports and adjusted the logic in each handler to inherit from the new base class.
- Ensured that all handlers implement the required methods: `name`, `lock_id`, `execute`, and `post_process`.
- Improved logging and event broadcasting in each handler to maintain functionality.
