+++ src/ai/backend/manager/sokovan/scheduler/handlers/base.py
@@ -1,0 +1,45 @@
+"""Base handler class for scheduler operations."""
+
+from abc import ABC, abstractmethod
+from typing import Optional
+
+from ai.backend.manager.defs import LockID
+from ai.backend.manager.sokovan.scheduler.results import ScheduleResult
+
+
+class SchedulerHandler(ABC):
+    """Base class for scheduler operation handlers."""
+
+    @classmethod
+    @abstractmethod
+    def name(cls) -> str:
+        """Get the name of the handler."""
+        raise NotImplementedError("Subclasses must implement name()")
+
+    @property
+    @abstractmethod
+    def lock_id(self) -> Optional[LockID]:
+        """Get the lock ID for this handler.
+
+        Returns:
+            LockID to acquire before execution, or None if no lock needed
+        """
+        raise NotImplementedError("Subclasses must implement lock_id")
+
+    @abstractmethod
+    async def execute(self) -> ScheduleResult:
+        """Execute the scheduling operation.
+
+        Returns:
+            Result of the scheduling operation
+        """
+        raise NotImplementedError("Subclasses must implement execute()")
+
+    @abstractmethod
+    async def post_process(self, result: ScheduleResult) -> None:
+        """Handle post-processing after the operation.
+
+        Args:
+            result: The result from execute()
+        """
+        raise NotImplementedError("Subclasses must implement post_process()")
