+++ src/ai/backend/manager/sokovan/scheduler/handlers/check_terminating_progress.py
@@ -1,0 +1,77 @@
+"""Handler for checking terminating progress of sessions."""
+
+import logging
+from typing import Optional
+
+from ai.backend.common.events.dispatcher import EventProducer
+from ai.backend.common.events.event_types.session.broadcast import (
+    SchedulingBroadcastEvent,
+)
+from ai.backend.common.events.types import AbstractBroadcastEvent
+from ai.backend.common.types import AccessKey
+from ai.backend.logging import BraceStyleAdapter
+from ai.backend.manager.data.session.types import SessionStatus
+from ai.backend.manager.defs import LockID
+from ai.backend.manager.repositories.scheduler.repository import SchedulerRepository
+from ai.backend.manager.scheduler.types import ScheduleType
+from ai.backend.manager.sokovan.scheduler.results import ScheduleResult
+from ai.backend.manager.sokovan.scheduler.scheduler import Scheduler
+from ai.backend.manager.sokovan.scheduling_controller import SchedulingController
+
+from .base import SchedulerHandler
+
+log = BraceStyleAdapter(logging.getLogger(__name__))
+
+
+class CheckTerminatingProgressHandler(SchedulerHandler):
+    """Handler for checking if TERMINATING sessions are ready to transition to TERMINATED."""
+
+    def __init__(
+        self,
+        scheduler: Scheduler,
+        scheduling_controller: SchedulingController,
+        event_producer: EventProducer,
+        repository: SchedulerRepository,
+    ):
+        self._scheduler = scheduler
+        self._scheduling_controller = scheduling_controller
+        self._event_producer = event_producer
+        self._repository = repository
+
+    @classmethod
+    def name(cls) -> str:
+        """Get the name of the handler."""
+        return "check-terminating-progress"
+
+    @property
+    def lock_id(self) -> Optional[LockID]:
+        """Lock for operations targeting TERMINATING sessions."""
+        return LockID.LOCKID_SOKOVAN_TARGET_TERMINATING
+
+    async def execute(self) -> ScheduleResult:
+        """Check if sessions in TERMINATING state have all kernels terminated."""
+        return await self._scheduler.check_terminating_progress()
+
+    async def post_process(self, result: ScheduleResult) -> None:
+        """Log the number of sessions that transitioned to TERMINATED and invalidate cache."""
+        await self._scheduling_controller.mark_scheduling_needed(ScheduleType.SCHEDULE)
+        log.info("{} sessions transitioned to TERMINATED state", len(result.scheduled_sessions))
+
+        # Invalidate cache for affected access keys
+        affected_keys: set[AccessKey] = {
+            event_data.access_key for event_data in result.scheduled_sessions
+        }
+        await self._repository.invalidate_keypair_concurrency_cache(list(affected_keys))
+        log.debug("Invalidated concurrency cache for {} access keys", len(affected_keys))
+
+        # Broadcast batch event for sessions that transitioned to TERMINATED
+        events: list[AbstractBroadcastEvent] = [
+            SchedulingBroadcastEvent(
+                session_id=event_data.session_id,
+                creation_id=event_data.creation_id,
+                status_transition=str(SessionStatus.TERMINATED),
+                reason=event_data.reason,
+            )
+            for event_data in result.scheduled_sessions
+        ]
+        await self._event_producer.broadcast_events_batch(events)
