+++ src/ai/backend/manager/sokovan/scheduler/handlers/schedule_sessions.py
@@ -1,0 +1,81 @@
+"""Handler for scheduling pending sessions."""
+
+import logging
+from typing import Optional
+
+from ai.backend.common.events.dispatcher import EventProducer
+from ai.backend.common.events.event_types.session.broadcast import (
+    SchedulingBroadcastEvent,
+)
+from ai.backend.common.events.types import AbstractBroadcastEvent
+from ai.backend.common.types import AccessKey
+from ai.backend.logging import BraceStyleAdapter
+from ai.backend.manager.data.session.types import SessionStatus
+from ai.backend.manager.defs import LockID
+from ai.backend.manager.repositories.scheduler.repository import SchedulerRepository
+from ai.backend.manager.scheduler.types import ScheduleType
+from ai.backend.manager.sokovan.scheduler.results import ScheduleResult
+from ai.backend.manager.sokovan.scheduler.scheduler import Scheduler
+from ai.backend.manager.sokovan.scheduling_controller import SchedulingController
+
+from .base import SchedulerHandler
+
+log = BraceStyleAdapter(logging.getLogger(__name__))
+
+
+class ScheduleSessionsHandler(SchedulerHandler):
+    """Handler for scheduling pending sessions."""
+
+    def __init__(
+        self,
+        scheduler: Scheduler,
+        scheduling_controller: SchedulingController,
+        event_producer: EventProducer,
+        repository: SchedulerRepository,
+    ):
+        self._scheduler = scheduler
+        self._scheduling_controller = scheduling_controller
+        self._event_producer = event_producer
+        self._repository = repository
+
+    @classmethod
+    def name(cls) -> str:
+        """Get the name of the handler."""
+        return "schedule-sessions"
+
+    @property
+    def lock_id(self) -> Optional[LockID]:
+        """Lock for operations targeting PENDING sessions."""
+        return LockID.LOCKID_SOKOVAN_TARGET_PENDING
+
+    async def execute(self) -> ScheduleResult:
+        """Schedule all pending sessions across scaling groups."""
+        log.trace("Scheduling sessions across all scaling groups")
+        return await self._scheduler.schedule_all_scaling_groups()
+
+    async def post_process(self, result: ScheduleResult) -> None:
+        """Request precondition check if sessions were scheduled and invalidate cache."""
+        # Request next phase first
+        await self._scheduling_controller.mark_scheduling_needed(ScheduleType.CHECK_PRECONDITION)
+        log.info(
+            "Scheduled {} sessions, requesting precondition check", len(result.scheduled_sessions)
+        )
+
+        # Invalidate cache for affected access keys
+        affected_keys: set[AccessKey] = {
+            session.access_key for session in result.scheduled_sessions
+        }
+        await self._repository.invalidate_keypair_concurrency_cache(list(affected_keys))
+        log.debug("Invalidated concurrency cache for {} access keys", len(affected_keys))
+
+        # Broadcast batch event for scheduled sessions
+        events: list[AbstractBroadcastEvent] = [
+            SchedulingBroadcastEvent(
+                session_id=event_data.session_id,
+                creation_id=event_data.creation_id,
+                status_transition=str(SessionStatus.SCHEDULED),
+                reason=event_data.reason,
+            )
+            for event_data in result.scheduled_sessions
+        ]
+        await self._event_producer.broadcast_events_batch(events)
